--- cypress/utils/plugins/security-analytics-dashboards-plugin/detectors.js
@@ -1,80 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-const { BACKEND_BASE_PATH } = require('../../base_constants');
-const { NODE_API } = require('./constants');
-
-Cypress.Commands.add('createDetector', (detectorJSON) => {
-  cy.request(
-    'POST',
-    `${BACKEND_BASE_PATH}${NODE_API.DETECTORS_BASE}`,
-    detectorJSON
-  );
-});
-
-Cypress.Commands.add(
-  'createAliasMappings',
-  (indexName, ruleTopic, aliasMappingsBody, partial = true) => {
-    const body = {
-      index_name: indexName,
-      rule_topic: ruleTopic,
-      partial: partial,
-      alias_mappings: aliasMappingsBody,
-    };
-    cy.request({
-      method: 'POST',
-      url: `${BACKEND_BASE_PATH}${NODE_API.MAPPINGS_BASE}`,
-      body: body,
-    });
-  }
-);
-
-Cypress.Commands.add('updateDetector', (detectorId, detectorJSON) => {
-  cy.request(
-    'PUT',
-    `${BACKEND_BASE_PATH}/${NODE_API.DETECTORS_BASE}/${detectorId}`,
-    detectorJSON
-  );
-});
-
-Cypress.Commands.add('deleteSAPDetector', (detectorName) => {
-  const body = {
-    from: 0,
-    size: 5000,
-    query: {
-      nested: {
-        path: 'detector',
-        query: {
-          bool: {
-            must: [{ match: { 'detector.name': detectorName } }],
-          },
-        },
-      },
-    },
-  };
-  cy.request({
-    method: 'POST',
-    url: `${BACKEND_BASE_PATH}${NODE_API.DETECTORS_BASE}/_search`,
-    failOnStatusCode: false,
-    body,
-  }).then((response) => {
-    if (response.status === 200) {
-      for (let hit of response.body.hits.hits) {
-        cy.request(
-          'DELETE',
-          `${BACKEND_BASE_PATH}${NODE_API.DETECTORS_BASE}/${hit._id}`
-        );
-      }
-    }
-  });
-});
-
-Cypress.Commands.add('deleteAllDetectors', () => {
-  cy.request({
-    method: 'DELETE',
-    url: `${BACKEND_BASE_PATH}/.opensearch-sap-detectors-config`,
-    failOnStatusCode: false,
-  });
-});
